adjusted existing code to work with new helpers and fixed bugs in array

+++ src/array/src/float_array.c
@@ -5,22 +5,22 @@ array_t *create_float_array()
     return create_array(sizeof(float));
 }
 
-array_error_t insert_float(array_t *arr, int index, float element)
+array_error_t insert_float(array_t *arr, size_t index, float element)
 {
     return insert(arr, index, &element);
 }
 
-array_error_t remove_float(array_t *arr, int index, float *output)
+array_error_t remove_float(array_t *arr, size_t index, float *output)
 {
     return remove_element(arr, index, output);
 }
 
-array_error_t get_float(array_t *arr, int index, float *output)
+array_error_t get_float(array_t *arr, size_t index, float *output)
 {
     return get(arr, index, output);
 }
 
-array_error_t set_float(array_t *arr, int index, float element)
+array_error_t set_float(array_t *arr, size_t index, float element)
 {
     return set(arr, index, &element);
 }
